Pre-populate checkbox questions in session flow
The session flow stores user responses in a session. The session can only be accessed
from controllers and views. This is a problem because the code to determine which
checkbox was selected is in a helper nested in a presenter, so we need to pass the
relevant values from the session to that helper.

We tried passing in the whole session, but that doesn't work, because even though the
session looks like a hash, it is an object and cannot be serialised.

The checkbox partial is used by both types of smart answer flow, not just sessions,
so not all of the values we need are available (i.e. params[:node_slug]) so the querying
of the session has been put inside a guard so the other smart-answers still work.