Use consistent spacing in email-link
We try to use spacing in one direction only, to help with a consistent spacing and predictable page flow. We aim a consistent 6 units margin after each component and no margins at the top. To preserve the distance between the previous govspeak instance and this one we add a standard separator. This can be taken out then the govspeak component will be updated to use a similar spacing technique.