Componentise escape-link
When the component was ported in this application from https://github.com/alphagov/govuk-coronavirus-find-support the component became a plain partial coupled with additional logic that generates the destroy link and the documentation for this component was missed. We convert this back to a component that can be previewed and tested in isolation.